Parse CSV file a single time
The server now accepts only JSON. The CSV files should be parsed in the front end.
