+++ dto/Cargo.toml
@@ -11,7 +11,6 @@ serde = { version = "1.0.218", features = ["derive"] }
 derive-getters = "0.5.0"
 chrono = { version = "0.4.40", features = ["serde"]}
 parameterized = "2.0.0"
-csv = "1.3.1"
 log = "0.4.26"
 
 [dev-dependencies]
