+++ dto/src/csv_member.rs
@@ -1,10 +1,7 @@
 use crate::member_to_check::MemberToCheck;
-use csv::{Reader, StringRecord};
 use derive_getters::Getters;
-use log::warn;
 use serde::{Deserialize, Serialize};
 use std::cmp::Ordering;
-use std::collections::BTreeSet;
 
 /// A [CsvMember] is a member whose been imported from a CSV file or string.
 /// It doesn't have much information, as we want to keep it simple
@@ -20,87 +17,15 @@ pub struct CsvMember {
 impl CsvMember {
     pub fn new(
         membership_num: String,
-        identity: Option<String>,
         name: Option<String>,
         first_name: Option<String>,
+        identity: Option<String>,
     ) -> Self {
         Self {
             membership_num,
-            identity,
             name,
             first_name,
-        }
-    }
-
-    /// Load members to check from a CSV-formatted String, such as:
-    /// `membership_num;name;firstname`
-    pub fn load_members_to_check_from_csv_string(
-        members_to_check: &str,
-    ) -> (BTreeSet<Self>, Vec<String>) {
-        let mut reader = csv::ReaderBuilder::new()
-            .delimiter(b';')
-            .has_headers(false)
-            .flexible(true)
-            .from_reader(members_to_check.as_bytes());
-
-        Self::load_members_to_check_from_csv(&mut reader)
-    }
-
-    /// Load members to check rom a CSV-formatted Reader, such as:
-    /// `membership_num;name;firstname`
-    fn load_members_to_check_from_csv<T>(reader: &mut Reader<T>) -> (BTreeSet<Self>, Vec<String>)
-    where
-        T: std::io::Read,
-    {
-        let mut members_to_check = BTreeSet::new();
-        let mut wrong_lines = vec![];
-
-        reader.records().for_each(|record| {
-            match Self::deserialize_member_to_check(record) {
-                Ok(member) => {
-                    members_to_check.insert(member);
-                }
-                Err(wrong_line) => {
-                    if let Some(wrong_line) = wrong_line {
-                        wrong_lines.push(wrong_line);
-                    }
-                }
-            };
-        });
-
-        (members_to_check, wrong_lines)
-    }
-
-    fn deserialize_member_to_check(
-        record: Result<StringRecord, csv::Error>,
-    ) -> Result<CsvMember, Option<String>> {
-        if let Ok(record) = record {
-            let fields_count = record.len();
-            if fields_count == 2 {
-                Ok(CsvMember::new(
-                    record.get(0).unwrap().to_owned(),
-                    Some(record.get(1).unwrap().to_owned()),
-                    None,
-                    None,
-                ))
-            } else if fields_count == 3 {
-                Ok(CsvMember::new(
-                    record.get(0).unwrap().to_owned(),
-                    None,
-                    Some(record.get(1).unwrap().to_owned()),
-                    Some(record.get(2).unwrap().to_owned()),
-                ))
-            } else {
-                Err(Some(record.iter().collect::<Vec<_>>().join(";")))
-            }
-        } else if let Err(error) = record {
-            warn!(
-                "Error while deserializing member to check [error: {:?}]",
-                error
-            );
-            Err(None)
-        } else {
-            Err(None)
+            identity,
         }
     }
 }
@@ -159,61 +84,6 @@ impl Ord for CsvMember {
 
 #[cfg(test)]
 mod tests {
-    mod load_members_to_check_from_csv_string {
-        use crate::csv_member::CsvMember;
-        use std::collections::BTreeSet;
-
-        #[test]
-        fn success_with_name_and_firstname() {
-            let membership_num = "123".to_owned();
-            let name = "Doe".to_owned();
-            let first_name = "John".to_owned();
-            let csv = format!("{membership_num};{name};{first_name}");
-            let result = CsvMember::load_members_to_check_from_csv_string(&csv);
-            assert_eq!(
-                (
-                    BTreeSet::from_iter(vec![CsvMember {
-                        membership_num,
-                        identity: None,
-                        name: Some(name),
-                        first_name: Some(first_name),
-                    }]),
-                    vec![]
-                ),
-                result
-            )
-        }
-
-        #[test]
-        fn success_with_identity() {
-            let membership_num = "123".to_owned();
-            let identity = "Doe John".to_owned();
-            let csv = format!("{membership_num};{identity}");
-            let result = CsvMember::load_members_to_check_from_csv_string(&csv);
-            assert_eq!(
-                (
-                    BTreeSet::from_iter(vec![CsvMember {
-                        membership_num,
-                        identity: Some(identity),
-                        name: None,
-                        first_name: None,
-                    }]),
-                    vec![]
-                ),
-                result
-            )
-        }
-
-        #[test]
-        fn fail_when_wrong_row() {
-            let membership_num = "123".to_owned();
-            let csv = format!("{membership_num}");
-            let result = CsvMember::load_members_to_check_from_csv_string(&csv);
-            let expected_result = (BTreeSet::new(), vec![csv]);
-            assert_eq!(expected_result, result)
-        }
-    }
-
     use crate::csv_member::CsvMember;
     use crate::member_to_check::MemberToCheck;
