+++ wasm/Cargo.toml
@@ -31,6 +31,7 @@ web-sys = { version = "0.3.77", features = ["DateTimeValue", "Document", "Docume
 serde = { version = "1.0.218", features = ["derive"] }
 log = "0.4.26"
 serde-json-wasm = "1.0.1"
+csv = "1.3.1"
 
 [dev-dependencies]
 wasm-bindgen-test = "0.3.50"
